Revert to previous form of all.yml
There is no good way, yet discovered, to ensure unneeded tasks are
not duplicated on a host. It is currently such that duplicate tasks
are not executed within a 'group', but if a host is in multiple
groups it is very possible for such a host to run duplicate tasks.