Allow the machine to be in any SELiunx state
Since most sites will already have it disabled/permissive this
allows Ansible to not have to manage it. It has been found that in
the rare case that it is enforcing, this boolean can be set to
allow httpd to properly proxy connections. No other errors related
to having it set to enforcing have been found.